Prevent marker-icons to cover each other in the maps.
Used some trigonometry to add a circular offset around the center. The
offset depends in the amount of topics that are displayed. I've also
consider storing the offsets OnInit so we don't have to calculate them
for each user/topics. I would like to change this in the future and use
the new spiderfier of GoogleMapsAPI v3 (currently on develop in angular-
google-maps)

Related links:
https://github.com/jawj/OverlappingMarkerSpiderfier
https://github.com/SebastianM/angular-google-maps/issues/1146

Closes #76